Create RequestButton.java
the program acts as a friend request handler, using Java Swing libraries to present a dialog box asking the user to accept or deny a friend request. After the user makes a choice, a notification is displayed indicating whether the request was accepted or denied. If no friend requests are received, a message notifies the user.